Verifier le foreign key

+++ TP/src/main/java/ma/ehei/gi4/tp/Services/CategoriesService.java
@@ -30,8 +30,15 @@ public class CategoriesService {
     }
 
     public void delete(Long id){
-        categorieRepository.deleteById(id);
+        if(categorieRepository.existsById(id)){
+            Category category=categorieRepository.findById(id).get();
+            if(category.getProduits().size()==0){
+                categorieRepository.deleteById(id);
+            }
+        }
+
     }
+
 
 
 }
